add support for mdspan pack/unpack
